Add documentation / chapters
I have been working on this quite a bit but I am sure there is a need for more updated or clarificiation but this pdf is supposed to further explain how you can use this repository or make a scrolling camera of your own! Will try to make it in to a readme.md later.